add separate code listings document
I've kept it in the same repo even though the main document doesn't use any of
the code listings' documents because the code listings are a part of the report.

To generate code listings for the project, you need to point the
`generate-code-listings` script to the project directory and run it, then render
`code-listings.tex` in the root directory.